added routing system and a visual frontend for lobby view
